backend/generate_report/requirements.txt
# Firebase
firebase-functions
firebase-admin

# Google Cloud AI for Gemini
google-cloud-aiplatform 